Fixed VC compiler warning in Hamamatsu.cpp
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/branches/micromanager1.3@1919 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Hamamatsu/Hamamatsu.cpp
@@ -1292,7 +1292,7 @@ int CHamamatsu::SnapImage()
 
    snapInProgress_ = true;
 
-   CDeviceUtils::SleepMs(dExp*1000.0);
+   CDeviceUtils::SleepMs((long) (dExp*1000.0));
 
    return DEVICE_OK;
 }
